requirements.txt
annotated-types          0.7.0
anyio                    3.7.1
certifi                  2025.1.31
charset-normalizer       3.4.1
click                    8.1.8
cmake                    4.0.0
exceptiongroup           1.2.2
faiss-cpu                1.10.0
fastapi                  0.104.1
filelock                 3.18.0
fsspec                   2025.3.2
h11                      0.14.0
huggingface-hub          0.16.4
idna                     3.10
Jinja2                   3.1.6
joblib                   1.4.2
lit                      18.1.8
MarkupSafe               3.0.2
mpmath                   1.3.0
networkx                 3.4.2
nltk                     3.9.1
numpy                    1.26.4
nvidia-cublas-cu11       11.10.3.66
nvidia-cuda-cupti-cu11   11.7.101
nvidia-cuda-nvrtc-cu11   11.7.99
nvidia-cuda-runtime-cu11 11.7.99
nvidia-cudnn-cu11        8.5.0.96
nvidia-cufft-cu11        10.9.0.58
nvidia-curand-cu11       10.2.10.91
nvidia-cusolver-cu11     11.4.0.1
nvidia-cusparse-cu11     11.7.4.91
nvidia-nccl-cu11         2.14.3
nvidia-nvtx-cu11         11.7.91
packaging                24.2
pillow                   11.2.1
pip                      23.0.1
pydantic                 2.4.2
pydantic_core            2.10.1
python-dotenv            1.0.0
PyYAML                   6.0.2
regex                    2024.11.6
requests                 2.32.3
safetensors              0.5.3
scikit-learn             1.6.1
scipy                    1.15.2
sentence-transformers    2.2.2
sentencepiece            0.2.0
setuptools               65.5.0
sniffio                  1.3.1
starlette                0.27.0
sympy                    1.13.3
threadpoolctl            3.6.0
tokenizers               0.13.3
torch                    2.0.1
torchvision              0.15.2
tqdm                     4.67.1
transformers             4.30.2
triton                   2.0.0
typing_extensions        4.13.2
urllib3                  2.4.0
uvicorn                  0.24.0
wheel                    0.45.1
